add <includeScopes> and <excludeScopes> configuration
The official Apache licensing howto says: "Dependencies which are not included in the distribution MUST NOT be added to LICENSE and NOTICE".
Therefore, I think it would be nice to be able to exclude test dependencies, because they are usually not bundled with a distribution.

This PR adds an configuration option like this:

<configuration>
    <includeScopes>
        <includeScopes>compile</includeScopes>
    </includeScopes>
</configuration>

You can either specify <includeScopes> to list only the scopes you want to have included in the NOTICE file,
or <excludeScopes> to list the scopes to be omitted from the NOTICE file.

See here for the quote above: http://www.apache.org/dev/licensing-howto.html#bundled-vs-non-bundled